Refactor recreate script further
